requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
clicksend-client @ git+https://github.com/ClickSend/clicksend-python.git@52baa3aebf70e4f39209fffbb16fe3199bf3b928
dj-database-url==0.5.0
Django==3.0.4
django-extensions==2.2.9
google-api-core==1.17.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gspread==3.3.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
lxml==4.5.0
numpy==1.18.3
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.3
pandas-gbq==0.13.1
protobuf==3.11.3
psycopg2==2.7.4
pyarrow==0.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.1.0
python-dateutil==2.8.1
pytrends==4.7.2
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
tzlocal==2.0.0
urllib3==1.25.8
whitenoise==5.0.1
